+++ article.json
@@ -1,11 +1,11 @@
 {
     "id": "4a71a08a-b31e-427e-84f3-127a1273ffd",
     "title": "Understanding Node.js File System Module",
-    "url": "https://levelup.gitconnected.com/understanding-node-js-file-system-module-b16da1e01949",
+    "url": "",
     "keywords": [1],
-    "modifiedAt": "06/24/2019",
-    "publishedAt": "06/24/2019",
+    "modifiedAt": "04/24/2019",
+    "publishedAt": "05/24/2019",
     "author": "Swathi Prasad",
-    "readMins": -10,
-    "source": "BLOG"
+    "readMins": 10,
+    "source": "NEWSPAPER"
   }
